requirements.txt
fastapi
uvicorn
python-multipart
pypdf
sentence-transformers
faiss-cpu
langchain
google-generativeai
python-dotenv
langchain
langchain-community
rank_bm25